requirements.txt
albucore==0.0.13
albumentations==1.4.10
annotated-types==0.7.0
anyio==4.7.0
astor==0.8.1
beautifulsoup4==4.12.3
boto3==1.35.79
botocore==1.35.79
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.11
decorator==5.1.1
fire==0.7.0
fonttools==4.55.3
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.36.1
imgaug==0.4.0
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
lmdb==1.5.1
lxml==5.3.0
matplotlib==3.9.3
networkx==3.4.2
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
packaging==24.2
paddleocr==2.9.1
paddlepaddle==2.6.2
pandas==2.2.3
pdf2image==1.17.0
pillow==11.0.0
protobuf==5.29.1
pyarrow==18.1.0
pyclipper==1.3.0.post6
pydantic==2.10.3
pydantic_core==2.27.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.10.1
requests==2.32.3
s3transfer==0.10.4
scikit-image==0.24.0
scikit-learn==1.6.0
scipy==1.14.1
shapely==2.0.6
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.9.20
tomli==2.2.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
